Fix Xamarin Incompatibility (#41)
When I try to convert the library to Xamarin, I have a small issue with access modifiers. I think C# is a little more strict than java on this, but the issue is that if an abstract method is protected, an override must be protected or private. It can't be made public. In the Transcoder library, DefaultDataSource.applyExtractor and DefaultDataSource.applyRetriever are protected, but all the overrides are public. This PR makes the subclass methods protected to remove the issue.