Enable multiple video tracks with no audio.
There's always a number of audio sources equal to the number of video sources (TranscoderOptions.java:131)
and getTrackFormat returns null when the video file had no audio track or vice versa.
So when getTrackFormat  returns null for all sources we can remove the associated track
type.

If only some of the data source files have no audio or video sources then it will still throw
an exception because we can't build a valid stream with missing parts.: it is either a full
track or no track.